create file "assignment2-1.py"

+++ assignment2-1.py
@@ -2,8 +2,8 @@ males = int(input("Enter number of males:"))
 females = int(input("Enter number of females:"))
 
 total = males + females
-male_percent = int((males/total) *100)
-female_percent = int((females/total) *100)
+male_percent = float((males/total) *100)
+female_percent = float((females/total) *100)
 
 print("The total number of students:\t", total)
 print("The number of males and females\t", males, "\t", females)
